feat:Entity, Repository, Service 1차 제작
fix:Oauth에서 네이버 부분에 nickname으로 되어있던 것을 name으로 변경
fix:전체적으로 잘못 구성되어있었던 Entity, Repository, Service 수정

+++ src/main/java/taehyeon/com/blog/repository/BlogRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import taehyeon.com.blog.entity.Blog;
 
 @Repository
-public interface BlogRepository extends JpaRepository<Blog, Integer> {
+public interface BlogRepository extends JpaRepository<Blog, Long> {
 }
